Added a couple more helper functions and removed their usage in timeline
